expose ok

+++ includes/struct.h
@@ -6,7 +6,7 @@
 /*   By: vquesnel <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/05/20 12:14:33 by vquesnel          #+#    #+#             */
-/*   Updated: 2016/05/24 16:32:18 by vquesnel         ###   ########.fr       */
+/*   Updated: 2016/05/24 20:24:14 by vquesnel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ typedef struct		s_env
 	double	planey;
 	clock_t	time;
 	clock_t	oldtime;
-	int		fps;
+	double		fps;
 	double	cam_x;
 	int		mapx;
 	int		mapy;
